docker-compose.yml

+++ src/main/resources/db/migration/V1__create-tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-    user_id SERIAL PRIMARY KEY,
+    user_id BIGSERIAL PRIMARY KEY,
     user_name VARCHAR(50) NOT NULL,
     email VARCHAR(100) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
@@ -9,33 +9,33 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS brand (
-    brand_id SERIAL PRIMARY KEY,
+    brand_id BIGSERIAL PRIMARY KEY,
     name VARCHAR(50) NOT NULL UNIQUE,
     country VARCHAR(255) CHECK (country IN ('Russia', 'Belarus',
     'USA', 'China', 'France', 'Italy', 'Korea'))
 );
 
   CREATE TABLE IF NOT EXISTS product (
-      product_id SERIAL PRIMARY KEY,
+      product_id BIGSERIAL PRIMARY KEY,
       name VARCHAR(50) NOT NULL,
       description TEXT,
       category VARCHAR(255) CHECK (category IN ('skincare', 'makeup', 'haircare', 'perfume',
       'nail_care', 'sun_care', 'men_cosmetics', 'organic', 'other')),
-      brand_id INT NOT NULL,
+      brand_id BIGINT NOT NULL,
       price NUMERIC(10,2) NOT NULL,
       CONSTRAINT fk_brand_product FOREIGN KEY (brand_id) REFERENCES brand(brand_id)
   );
 
 CREATE TABLE IF NOT EXISTS cart (
-    cart_id SERIAL PRIMARY KEY,
-    user_id INT NOT NULL UNIQUE,
+    cart_id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT NOT NULL UNIQUE,
     total_price NUMERIC(10,2),
     CONSTRAINT fk_user_cart FOREIGN KEY (user_id) REFERENCES users(user_id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS orders (
-    order_id SERIAL PRIMARY KEY,
-    user_id INT NOT NULL,
+    order_id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT NOT NULL,
     total_price NUMERIC(10,2) NOT NULL,
     date DATE NOT NULL,
     status VARCHAR(255) NOT NULL CHECK (status IN ('completed',
@@ -44,33 +44,33 @@ CREATE TABLE IF NOT EXISTS orders (
 );
 
 CREATE TABLE IF NOT EXISTS order_item (
-    order_item_id SERIAL PRIMARY KEY,
-    order_id INT NOT NULL,
-    product_id INT NOT NULL,
-    quantity INT NOT NULL,
+    order_item_id BIGSERIAL PRIMARY KEY,
+    order_id BIGINT NOT NULL,
+    product_id BIGINT NOT NULL,
+    quantity BIGINT NOT NULL,
     CONSTRAINT fk_order_orderitem FOREIGN KEY (order_id) REFERENCES orders(order_id) ON DELETE CASCADE,
     CONSTRAINT fk_product_orderitem FOREIGN KEY (product_id) REFERENCES product(product_id)
 );
 
 CREATE TABLE IF NOT EXISTS cart_item (
-    cart_item_id SERIAL PRIMARY KEY,
-    cart_id INT NOT NULL,
-    product_id INT NOT NULL,
-    quantity INT NOT NULL,
+    cart_item_id BIGSERIAL PRIMARY KEY,
+    cart_id BIGINT NOT NULL,
+    product_id BIGINT NOT NULL,
+    quantity BIGINT NOT NULL,
     CONSTRAINT fk_cart_cartitem FOREIGN KEY (cart_id) REFERENCES cart(cart_id) ON DELETE CASCADE,
     CONSTRAINT fk_product_cartitem FOREIGN KEY (product_id) REFERENCES product(product_id)
 );
 
 CREATE TABLE IF NOT EXISTS refresh_token (
-    token_id SERIAL PRIMARY KEY,
+    token_id BIGSERIAL PRIMARY KEY,
     token VARCHAR(255) NOT NULL,
-    user_id INT NOT NULL,
+    user_id BIGINT NOT NULL,
     expiry_date TIMESTAMP NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(user_id) ON DELETE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS black_list (
-    token_id SERIAL PRIMARY KEY,
+    token_id BIGSERIAL PRIMARY KEY,
     token VARCHAR(255) NOT NULL,
     expiry_date TIMESTAMP NOT NULL
 );
